Cleanup modular features and use objects rather than Double.NaN

+++ src/main/java/io/github/mzmine/datamodel/features/Feature.java
@@ -45,42 +45,42 @@ public interface Feature {
   /**
    * This method returns raw M/Z value of the feature
    */
-  double getMZ();
+  Double getMZ();
 
   /**
    * Sets raw M/Z value of the feature
    */
-  void setMZ(double mz);
+  void setMZ(Double mz);
 
   /**
    * This method returns raw retention time of the feature in minutes
    */
-  float getRT();
+  Float getRT();
 
   /**
    * Sets retention time of the feature
    */
-  void setRT(float rt);
+  void setRT(Float rt);
 
   /**
    * This method returns the raw height of the feature
    */
-  float getHeight();
+  Float getHeight();
 
   /**
    * Sets height of the feature
    */
-  void setHeight(float height);
+  void setHeight(Float height);
 
   /**
    * This method returns the raw area of the feature
    */
-  float getArea();
+  Float getArea();
 
   /**
    * Sets area of the feature
    */
-  void setArea(float area);
+  void setArea(Float area);
 
   /**
    * Returns raw data file where this feature is present
@@ -114,7 +114,7 @@ public interface Feature {
    * @return
    */
   @Nullable
-  default float getRetentionTimeAtIndex(int i) {
+  default Float getRetentionTimeAtIndex(int i) {
     List<Scan> scans = getScanNumbers();
     return scans == null ? null : scans.get(i).getRetentionTime();
   }
@@ -280,32 +280,32 @@ public interface Feature {
   /**
    * This method returns the full width at half maximum (FWHM) of the feature
    */
-  float getFWHM();
+  Float getFWHM();
 
   /**
    * Sets the full width at half maximum (FWHM)
    */
-  void setFWHM(double fwhm);
+  void setFWHM(Float fwhm);
 
   /**
    * This method returns the tailing factor of the feature
    */
-  float getTailingFactor();
+  Float getTailingFactor();
 
   /**
    * Sets the tailing factor
    */
-  void setTailingFactor(double tf);
+  void setTailingFactor(Float tf);
 
   /**
    * This method returns the asymmetry factor of the feature
    */
-  float getAsymmetryFactor();
+  Float getAsymmetryFactor();
 
   /**
    * Sets the asymmetry factor
    */
-  void setAsymmetryFactor(double af);
+  void setAsymmetryFactor(Float af);
 
   // dulab Edit
   void outputChromToFile();
